# docs/swagger.yaml
basePath: /api
...
host: localhost:5000
info:
  contact: {}
  description: 兽剧内容聚合平台后端服务 API 文档（neo-server，行为对齐 Express 版）
  title: 兽剧聚合平台 API
  version: 1.0.0
paths:
...
  /user-sessions/{id}:
    delete:
      parameters:
      - description: 会话 ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: 下线指定会话
      tags:
      - 用户会话
  /user-sessions/{id}/name:
    put:
      parameters:
      - description: 会话 ID
        in: path
        name: id
        required: true
        type: string
      - description: deviceName
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: 重命名会话
      tags:
      - 用户会话
  /user-sessions/admin/{id}:
    delete:
      parameters:
      - description: 会话 ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: 管理员下线会话
      tags:
      - 用户会话
  /user-sessions/admin/user/{userId}/all:
    delete:
      parameters:
      - description: 用户 ID
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: 管理员下线用户全部会话
      tags:
      - 用户会话
  /user-sessions/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - bearerAuth: []
      summary: 全部会话（超管）
      tags:
      - 用户会话
  /user-sessions/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: screenWidth/screenHeight/language
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "200":
          description: sessionId
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: 创建用户会话（记录设备）
      tags:
      - 用户会话
  /user-sessions/heartbeat:
    post:
      responses:
        "200":
          description: ok
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - bearerAuth: []
      summary: 会话心跳
      tags:
      - 用户会话
  /user-sessions/my:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - bearerAuth: []
      summary: 我的会话列表
      tags:
      - 用户会话
  /user-sessions/my/all:
    delete:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: 下线其他所有设备
      tags:
      - 用户会话
...
securityDefinitions:
  bearerAuth:
    description: 输入 JWT access token（不含 Bearer 前缀；亦可通过 accessToken cookie）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"